requirements/dev.txt
bpython
django-debug-toolbar
django-debugtools
django-model-info
django-stubs
django-pdb
django-anonymous
django-stubs-ext
djangorestframework-stubs
djangorestframework-stubs[compatible-mypy]
django-debug-permissions
django-request-viewer
django-silk
djipsum
gprof2dot
faker
factory_boy
inquirer
phone-gen
pip-chill
pipdeptree
prettyprinter
#pygraphviz
pyparsing
pdbpp
pydot
docutils
PyPika
remote-pdb
tabulate
rich
termcolor
Werkzeug
pyOpenSSL